+++ index.php
@@ -3,9 +3,9 @@
  * Plugin Name:       CodeSoup ACF Admin Categories
  * Plugin URI:        https://github.com/code-soup/acf-admin-categories
  * Description:       A WordPress plugin that adds category organization to Advanced Custom Fields (ACF) field groups for better management and organization.
- * Version:           1.0.1
+ * Version:           1.0.2
  * Requires at least: 6.0
- * Requires PHP:      8.0
+ * Requires PHP:      8.1
  * Author:            Code Soup
  * Author URI:        https://www.codesoup.co
  * License:           GPL v3 or later